==> 12-ingestao-dados-servidores-afastamento.ipynb <==
import pandas as pd

!bzcat origem/202012_Afastamentos.csv.bz2 | head

df = pd.read_csv('origem/202012_Afastamentos.csv.bz2', sep=';', 
                        usecols=['Id_SERVIDOR_PORTAL', 'NOME'], index_col=0)
df.head()

df.index

df.drop_duplicates(inplace=True)

df.to_csv('afastamentos.csv', sep=';')

!rm -f afastamentos.csv.bz2 && bzip2 -9 afastamentos.csv

!bzcat afastamentos.csv.bz2 | head

!mv afastamentos.csv.bz2 dados/
