==> 13-ingestao-dados-servidores-cadastro.ipynb <==
import pandas as pd

!bzcat origem/202012_Cadastro.csv.bz2 | head -3

df = pd.read_csv('origem/202012_Cadastro.csv.bz2', sep=';', 
                        usecols=['Id_SERVIDOR_PORTAL', 'NOME'], index_col=0)
df.head()

df.index

df.drop_duplicates(inplace=True)

df.to_csv('cadastro.csv', sep=';')

!rm -f cadastro.csv.bz2 && bzip2 -9 cadastro.csv

!bzcat cadastro.csv.bz2 | head

!mv cadastro.csv.bz2 dados/
