==> 21-analise-servidores-nomes.ipynb <==
import pandas as pd
import re

#arquivo = 'dados/afastamentos'
arquivo = 'dados/cadastro'

df = pd.read_csv(arquivo + '.csv.bz2', sep=';',
                 usecols=['Id_SERVIDOR_PORTAL', 'NOME'], index_col=0)

df.head(10)

def remover_preposicoes(str):
    return re.sub(" D[A|E|O]S? ", " ", str)

exemplos = [
    "BELCHIOR DA CUNHA NETO",
    "ABADIO ALVES DE LIMA FILHO",
    "ALBINO GOMES DOS SANTOS JUNIOR"
]

for ex in exemplos:
    print(ex, "=>", remover_preposicoes(ex))

df['NOME2'] = df['NOME'].apply(remover_preposicoes)
df.head(30)

def obter_primeiro_nome(str):
    return re.split("\s", str)[0]

for ex in exemplos:
    print(ex, "=>", obter_primeiro_nome(ex))

def obter_ultimo_nome(str):
    nomes = re.split("\s", str)
    ultimo = nomes[-1]
    if ultimo not in ["FILHO", "JUNIOR", "NETO", "SOBRINHO", "JR", "SEGUNDO", "TERCEIRO"]:
        return ultimo
    else:
        return nomes[-2]

for ex in exemplos:
    print(ex, "=>", obter_ultimo_nome(ex))

df['PNOME'] = df['NOME2'].apply(obter_primeiro_nome)
df['UNOME'] = df['NOME2'].apply(obter_ultimo_nome)
df.head(30)

## Nomes

maiores_nomes = df[['PNOME', 'UNOME']].\
    groupby(['PNOME']).count().sort_values(['UNOME'], ascending=False)
maiores_nomes.rename(columns={'UNOME': 'QTDE'}, inplace=True)

qtd_total = maiores_nomes['QTDE'].sum()
qtd_total

maiores_nomes['PERC'] = maiores_nomes['QTDE'].apply(lambda x: x / qtd_total * 100.0)
maiores_nomes.head(30)

## Sobrenomes

maiores_sobrenomes = df[['PNOME', 'UNOME']].\
    groupby(['UNOME']).count().sort_values(['PNOME'], ascending=False)
maiores_sobrenomes.rename(columns={'PNOME': 'QTDE'}, inplace=True)

maiores_sobrenomes['PERC'] = maiores_sobrenomes['QTDE'].apply(lambda x: x / qtd_total * 100.0)
maiores_sobrenomes.head(30)

### Geração dos arquivos resultantes

maiores_nomes.to_csv(arquivo + '-nomes.csv', float_format='%.4f')
maiores_sobrenomes.to_csv(arquivo + '-sobrenomes.csv', float_format='%.4f')
