==> 31-rotulagem-genero-nome.ipynb <==
import pandas as pd
import numpy as np

arquivo = 'dados/afastamentos'
#arquivo = 'dados/cadastro'

df = pd.read_csv(arquivo + '-nomes.csv', index_col=0,
                 dtype={'QTDE': 'uint32', 'PERC': np.float32})

df.head(30)

df['SEXO'] = None
#df['LETRAS'] = 0
df['REV'] = df.index
df['REV'] = df['REV'].apply(lambda x: x[::-1])
#df['LETRAS'] = df['REV'].apply(lambda x: len(x))
df.head(30)

df.dtypes

### 1) Masculino se terminar com O, Feminino se terminar com A

df.loc[df['REV'].str.startswith('O'), 'SEXO'] = 'M'
df.loc[df['REV'].str.startswith('A'), 'SEXO'] = 'F'

### 2) Masculino se terminar com R, X ou L

df.loc[df['REV'].str.startswith('R'), 'SEXO'] = 'M'
df.loc[df['REV'].str.startswith('X'), 'SEXO'] = 'M'
df.loc[df['REV'].str.startswith('L'), 'SEXO'] = 'M'

### 3) Masculino se terminar com ...

sufixos = [
"DRE",
"UIS",
"QUE",
"ERT",
"ERG"
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'SEXO'] = 'M'

### 4) Masculino se terminar com ...

sufixos = [
"OS",
"ON",
"AS",
"AZ",
"US",
"ES",
"PE",
"PH",
"IM",
"NS",
"UE",
"EI",
"EY",
"EU",
"AU",
"AN",
"AC",
"IC",
"ICK",
"IK"
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'SEXO'] = 'M'

### 5) Feminino se terminar com ...

sufixos = [
"ELE",
"ELLE",
"LLE",
"ETE",
"ETTE",
"ETI",
"ETTI",
"ETH",
"ETHE",
"ISE",
"IZE",
"ITE"
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'SEXO'] = 'F'

### 6) Feminino se terminar com ...

sufixos = [
"NE",
"LI",
"LY",
"LYN",
"CE",
"EN",
"DE",
"SE",
"SY",
"ANI",
"AIS",
"AH",
"ENI",
"IE",
"IRE"
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'SEXO'] = 'F'

### 8) Nomes Masculinos ainda não classificados ou reclassificados

nomes = [
"JOSE",
"JORGE",
"LUIZ",
]

with open('rotulagem/nomes-masculinos.txt', 'w') as f:
    for nome in nomes:
        f.write("%s\n" % nome)

nomes = []
with open('rotulagem/nomes-masculinos.txt', 'r') as f:
    nomes = [nome.rstrip() for nome in f.readlines()]

for nome in nomes:
    df.loc[df['REV'] == nome[::-1], 'SEXO'] = 'M'

### 9) Nomes Femininos ainda não classificados ou reclassificados

nomes = [
"RAQUEL",
"ISABEL",
"LEONOR"
]

with open('rotulagem/nomes-femininos.txt', 'w') as f:
    for nome in nomes:
        f.write("%s\n" % nome)

nomes = []
with open('rotulagem/nomes-femininos.txt', 'r') as f:
    nomes = [nome.rstrip() for nome in f.readlines()]

for nome in nomes:
    df.loc[df['REV'] == nome[::-1], 'SEXO'] = 'F'

### Casos indefinidos

df[df['SEXO'].isnull()].head(30)

### Testes com outros casos

df = df[df['REV'].str.startswith('IA')]

df[df['SEXO'].isnull()].head(100)

df.describe(include=['O'])

df.head(200)

### Testes com outros casos

df[df['REV'].str.startswith('EZI')]

#df.loc['PHILIP']

## Últimas transformações

df.loc[df['SEXO'].isnull(), 'SEXO'] = 'X'

## Análise das quantidades

df2 = df[['QTDE', 'SEXO']].groupby(['SEXO']).count()
total = df2['QTDE'].sum()
df2['PERC'] = df2['QTDE'].apply(lambda x: int(x / total * 10000) / 100)
df2

## Geração do arquivo final

df[['SEXO', 'QTDE']].to_csv(arquivo + '-nomes-sexo.csv', header=True)
