==> 32-rotulagem-pais-sobrenome.ipynb <==
import pandas as pd
import numpy as np

#BASE = 'dados/afastamentos'
BASE = 'dados/cadastro'

df = pd.read_csv(BASE + '-sobrenomes.csv', index_col=0,
                 dtype={'QTDE': 'uint32', 'PERC': np.float32})

df.drop(df[df.index.isna()].index, inplace=True)
#df = df.loc[~df.index.isna()]

df.head()

df.dtypes

#df['LETRAS'] = 0
df['REV'] = df.index
df['REV'] = df['REV'].apply(lambda x: x[::-1])
#df['LETRAS'] = df['REV'].apply(lambda x: len(x))
df['PAIS'] = None
df.head()

df.dtypes

maior = df['LETRAS'].max()
maior

maior = 16 # fixado para o maior arquivo

### Constantes

PORTUGAL = 'PT'
ESPANHA = 'ES'
ITALIA = 'IT'
ALEMANHA = 'DE'
FRANCA = 'FR'
JAPAO = 'JP'
POLONIA = 'PL'
RUSSIA = 'RU'
UCRANIA = 'UA'

### 1) Espanhol se terminar com ...

sufixos = [
"EZ",
"LLO",
"LLON",
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'PAIS'] = ESPANHA

### 2.1) Italiano se terminar com ...

sufixos = [
"CHIA",
"ETTA",
"GLIA",
"INI",
"IERI",
"ELLA",
"ELLI",
"ELLO",
"ETTI",
"ETTO",
"ONI",
"ONO",
"OTTI",
"OTTO",
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'PAIS'] = ITALIA

### 2.2) Italiano se contiver ...

termos = [
"CCH",
"GH",
"ZZ",
]

for termo in termos:
    df.loc[df['REV'].str.contains(termo), 'PAIS'] = ITALIA

### 3) Francês se terminar com ...

sufixos = [
"AMP",
"ARD",
"AUD",
"AUX",
"EAU",
"EAULT",
"ET",
"EU",
"EUX",
"IEU",
"ILLE",
"IER",
"OT",
"OIS",
"OIT",
"OUX",
"FORT",
"GNE",
"MONT"
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'PAIS'] = FRANCA

### 4.1) Alemão se terminar com ...

sufixos = [
"BACH",
"TNER",
"STEIN",
"MANN",
"ULLER",
"HOFF",
"WALD",
"BAUM",
"BERG",
"BERGER",
"HAUS",
"MAYER",
"GARTEN",
"AUER",
"KAMP",
"HARDT",
"RINGER",
"STIEL",
"HEIM",
"FELD",
"FELDT",
"HORN",
"BURG",
"FRIED",
"KOPF",
"REICH",
"MEISTER",
"MEYER",
"BLATT",
"INGER",
"MACHER",
"HOLZ"
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'PAIS'] = ALEMANHA

### 4.2) Alemão se iniciar com ...

prefixos = [
"HOFF",
"HART",
"BERG",
"SCHU",
"BECKEN",
]

for pref in prefixos:
    df.loc[df.index.str.startswith(pref), 'PAIS'] = ALEMANHA

### 5.1) Japonês se terminar com ...

sufixos = [
"AKA",
"AKI",
"ASHI",
"ATSU",
"AWA",
"HARA",
"HARU",
"HIDA",
"HIMA",
"HIRO",
"KITA",
"OKA",
"OSHI",
"MARU",
"MORI",
"MOTO",
"MURA",
"NAGA",
"NAKA",
"RAYA",
"UCHI",
"UKA",
"UMA",
"YAMA",
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'PAIS'] = JAPAO

### 5.2) Japonês se iniciar com ...

prefixos = [
"MATSU",
"FUJI",
"KOBA",
"YAMA",
"SAKI",
"SAKA",
"NAKA",
"TAKA",
"TAKE",
"YOSHI",
"NISHI",
"UEM",
"OKU",
"OKI",
"YOKO",
"MIYA",
"NISI",
"MORI",
"TASHI",
]

for pref in prefixos:
    df.loc[df.index.str.startswith(pref), 'PAIS'] = JAPAO

### 6) Polonês se terminar com ...

sufixos = [
"EVSKI",
"EWSKI",
"INSKI",
"INSKY",
"OVSKI",
"OWSKI",
"VICZ",
"WICZ",
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'PAIS'] = POLONIA

### 7) Russo se terminar com ...

sufixos = [
"VANOV",
"VITCH",
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'PAIS'] = RUSSIA

### 8) Ucraniano se terminar com ...

sufixos = [
"CZUK",
"CZYK",
]

for suf in sufixos:
    df.loc[df['REV'].str.startswith(suf[::-1]), 'PAIS'] = UCRANIA

IT:
.XXX
.XX.

DE:
.XXX

PL:
.XXX

RU:
.XXX

UA:
.XXX

### 8) Sobrenomes Espanhóis ainda não classificados ou reclassificados

nomes = []
with open('sobrenomes-espanhois.txt', 'r') as f:
    nomes = [nome.rstrip() for nome in f.readlines()]

for nome in nomes:
    df.loc[df['REV'] == nome[::-1], 'PAIS'] = ESPANHA

### 9) Sobrenomes Portugueses ainda não classificados ou reclassificados

nomes = []
with open('rotulagem/sobrenomes-portugueses.txt', 'r') as f:
    nomes = [nome.rstrip() for nome in f.readlines()]

for nome in nomes:
    df.loc[df['REV'] == nome[::-1], 'PAIS'] = PORTUGAL

## Testes individuais

df.loc['MELO']

### Casos indefinidos

df[df['PAIS'].isnull()].head()

### Análise dos resultados por país

df[df['PAIS'] == PORTUGAL].head(10)

df[df['PAIS'] == ESPANHA].head(10)

df[df['PAIS'] == ITALIA].head(10)

df[df['PAIS'] == ALEMANHA].head(10)

df[df['PAIS'] == FRANCA].head(10)

df[df['PAIS'] == JAPAO].head(10)

df[df['PAIS'] == POLONIA].head(10)

df[df['PAIS'] == RUSSIA].head(10)

df[df['PAIS'] == UCRANIA].head(10)

### Geração do arquivo de depuração (não identificados)

df2 = (df[df['PAIS'] == 'XX'])[['QTDE', 'REV']].sort_values(['REV', 'QTDE'])[['QTDE']]
#df2.head(30)
df2.to_csv(BASE + '-sobrenomes-pais-XX.csv')

### Geração do arquivo final

df.loc[df['PAIS'].isnull(), 'PAIS'] = 'XX'

df[['QTDE', 'PAIS']].groupby(['PAIS']).count().sort_values(['QTDE'], ascending=False)

df['PAIS'].to_csv(BASE + '-sobrenomes-pais.csv', float_format='%.4f')
