==> 41-preparo-atributos-genero-nome.ipynb <==
import pandas as pd
import numpy as np

## Carga dos dados de entrada

# definir arquivo CSV de entrada
#arquivo = 'dados/afastamentos-nomes-sexo.csv'
arquivo = 'dados/cadastro-nomes-sexo.csv'
#arquivo = 'extra/nome_sexo_pf.csv.bz2'

partes = arquivo.split('.')
prefixo_arquivo = partes[0]
#sufixo_arquivo = '.'.join(partes[1:])
novo_arquivo = prefixo_arquivo + '-preparado'
novo_arquivo

df = pd.read_csv(arquivo, index_col=0)
df.head()

# remover valores nulos
df = df.loc[~df.index.isna()]
df.drop(df[df.index.isnull()].index, inplace=True)

total = df[df.index.isnull()].shape[0]
print("linhas nulas:", total)

# renomear índice e coluna
df.index.rename('NOME', inplace=True)
df.rename(columns={
    df.columns[0]: 'SEXO',
    df.columns[1]: 'QTDE'
}, inplace=True)

df.head()

## Transformações nos dados

df['SEXO'] = df['SEXO'].map({'F': 0, 'M': 1, 'X': 9}).astype('uint8')

# remover sexo diferente de 0 ou 1
df.drop(df[df['SEXO'] > 1].index, inplace=True)

# converter para tipo de inteiro menor
df['QTDE'] = df['QTDE'].astype('uint32')
#np.iinfo(np.dtype('uint32'))

# criar reverso do nome
df['REV'] = df.index
df['REV'] = df['REV'].apply(lambda x: x[::-1])

df.head()

df['LETRAS'] = 0
df['LETRAS'] = df['REV'].apply(lambda x: len(x))
maior = df['LETRAS'].max()
maior

MAIOR_QTDE_LETRAS = 16 # fixado para o maior arquivo

for i in range(MAIOR_QTDE_LETRAS):
    df['L' + str(i)] = df['REV'].apply(
        lambda x: ord(x[i]) - 64 if len(x) > i else 0
    ).astype('uint8')

for i in range(MAIOR_QTDE_LETRAS):
    col = 'L' + str(i)
    #df[col] = df['REV'].apply(lambda x: x[i] if len(x) > i else ' ')
    #df.loc[df[col].isnull(), col] = ' '
    df[col] = df[col].apply(lambda x: ord(x) - 64 if x != ' ' else 0).astype('uint8')

# remover coluna reversa
df.drop(['REV'], axis=1, inplace=True)

df.dtypes

## Análise dos resultados

### Amostra dos dados

df.head(10)

print("total de linhas:", df.shape[0])

### Quantidades por sexo

df2 = df[['SEXO', 'L0']].groupby(by=['SEXO']).count().rename({'L0': 'CONT'}, axis=1)
total = df2['CONT'].sum()
df2['PERC'] = df2['CONT'].apply(lambda x: int(x / total * 10000) / 100)
df2

## Gravação do arquivo final

df.to_csv(novo_arquivo + '.csv.bz2', compression='bz2')

df.to_pickle(novo_arquivo + '.pkl.bz2')
