==> 53-modelo-genero-nome-hyperopt.ipynb <==
import pandas as pd
import numpy as np

# definir arquivo Pickle de entrada
arquivo = 'afastamentos-nomes-sexo-preparado.pkl.bz2'
#arquivo = 'cadastro-nomes-sexo-preparado.pkl.bz2'
#arquivo = 'extra/nome_sexo_pf05.csv'

df = pd.read_pickle(arquivo)
df.head()

df2 = df.copy()
df2['TEMP'] = df2[['SEXO', 'QTDE']].apply(lambda x: [x[0]] * x[1], axis=1)
df3 = df2.explode('TEMP').drop(['TEMP', 'QTDE'], axis=1) #.reset_index()

# usar os dados expandidos invés dos agregados
df0 = df # guardar o dataframe original
df = df3 # apontar para o novo dataframe

X = df.drop(['SEXO'], axis=1) # tudo, exceto a coluna alvo
y = df['SEXO'] # apenas a coluna alvo
print('Forma dos dados:', X.shape, y.shape)

# dividir dados em treino e teste

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.33, random_state=42)

print("Shapes:", X_train.shape, X_test.shape, y_train.shape, y_test.shape)

# treinar com todos os dados disponíveis

X_train, y_train = X, y

print("Shapes:", X_train.shape, y_train.shape)

https://github.com/hyperopt/hyperopt-sklearn

#pip install git+https://github.com/hyperopt/hyperopt-sklearn

#import hpsklearn as hs
from hpsklearn import HyperoptEstimator, any_classifier, any_preprocessing, any_sparse_classifier
from hyperopt import tpe

# Instantiate a HyperoptEstimator with the search space and number of evaluations
automl = HyperoptEstimator(
    classifier=any_sparse_classifier('my_clf'),
    preprocessing=[],
    algo=tpe.suggest,
    max_evals=5,
    trial_timeout=2*60,
    n_jobs=2
)
automl

type(automl)

# Search the hyperparameter space based on the data
automl.fit(X_train, y_train)

# Show the results
print(automl.score(X_test, y_test))

print(automl.best_model())

# carregar dados de teste: definir arquivo Pickle de entrada
#arquivo = 'afastamentos-nomes-sexo-preparado.pkl.bz2'
arquivo = 'cadastro-nomes-sexo-preparado.pkl.bz2'
#arquivo = 'extra/nome_sexo_pf05.csv'

df2 = pd.read_pickle(arquivo)
df2.head()

X_test = df2.drop(['SEXO'], axis=1) # tudo, exceto a coluna alvo
y_test = df2['SEXO'] # apenas a coluna alvo

y_pred = automl.predict(X_test)
y_pred

from sklearn.metrics import accuracy_score

print("Accuracy score: %.2f %%" % (accuracy_score(y_test, y_pred) * 100))

from sklearn.metrics import confusion_matrix

print("Confusion matrix:\n", confusion_matrix(y_test, y_pred))
