==> 59-avaliacao-individual-modelo-sklearn.ipynb <==
import numpy as np
import re

## Carregar modelo treinado

# definir arquivo Joblib de entrada
modelo = 'modelos/modelo-genero-nome.joblib.bz2' # sklearn (best)
#modelo = 'modelos/modelo-genero-nome-ABDT.joblib.bz2' # AdaBoost
#modelo = 'modelos/modelo-genero-nome-RF.joblib.bz2' # RF
#modelo = 'modelos/modelo-genero-nome-XGB.joblib.bz2' # XGBoost
#modelo = 'modelos/modelo-genero-nome-XGB-npf.joblib.bz2' # XGBoost (+dados)
#modelo = 'modelos/modelo-genero-nome-autoskl.joblib.bz2' # autosklearn

# carregar um modelo previamente gravado em disco
from joblib import load
model = load(modelo)
model

## Avaliação de exemplos individuais

def obter_primeiro_nome(str):
    if str:
        return re.split("\s", str)[0].upper()

MAIOR_QTDE_LETRAS = 16

def obter_array_letras(str):
    if not str:
        return None
    rev = str[::-1]
    anome = []
    for i in range(MAIOR_QTDE_LETRAS):
        x = ord(rev[i]) - 64 if len(rev) > i else 0
        anome.append(x)
    return anome

### Nulo ou vazio

obter_array_letras(obter_primeiro_nome(None))

obter_array_letras(obter_primeiro_nome(""))

### Feminino

nome = "Patricia da Silva"
anome1 = obter_array_letras(obter_primeiro_nome(nome))
print(anome1)
model.predict(np.asarray([anome1]))[0]

nome = "LEONOR BRITO"
model.predict(np.asarray([obter_array_letras(obter_primeiro_nome(nome))]))[0]

### Masculino

nome = "Arlindo de Souza"
anome2 = obter_array_letras(obter_primeiro_nome(nome))
print(anome2)
model.predict(np.asarray([anome2]))[0]

nome = "DJALMA DOS SANTOS"
model.predict(np.asarray([obter_array_letras(obter_primeiro_nome(nome))]))[0]
