==> 61-carga-dados-nome-sexo-pyspark.ipynb <==
## Overview

This notebook will create and query a table or DataFrame that you uploaded to DBFS. [DBFS](https://docs.databricks.com/user-guide/dbfs-databricks-file-system.html) is a Databricks File System that allows you to store data for querying inside of Databricks. This notebook assumes that you have a file already inside of DBFS that you would like to read from.

This notebook is written in **Python** so the default cell type is Python. However, you can use different languages by using the `%LANGUAGE` syntax. Python, Scala, SQL, and R are all supported.

https://docs.databricks.com/spark/latest/dataframes-datasets/introduction-to-dataframes-python.html

# File location and type
file_location = "/FileStore/tables/cadastro_nomes_sexo.csv"
file_type = "csv"

# CSV options
infer_schema = "false"
first_row_is_header = "true"
delimiter = ","

# The applied options are for CSV files. For other file types, these will be ignored.
df = spark.read.format(file_type) \
  .option("inferSchema", infer_schema) \
  .option("header", first_row_is_header) \
  .option("sep", delimiter) \
  .load(file_location)

display(df)

# Create a view or table

temp_table_name = "cadastro_nomes_sexo"

df.createOrReplaceTempView(temp_table_name)

%sql

/* Query the created temp table in a SQL cell */

select * from `cadastro_nomes_sexo` limit 100

# With this registered as a temp view, it will only be available to this particular notebook. If you'd like other users to be able to query this table, you can also create a table from the DataFrame.
# Once saved, this table will persist across cluster restarts as well as allow various users across different notebooks to query this data.
# To do so, choose your table name and uncomment the bottom line.

permanent_table_name = "cadastro_nomes_sexo"

df.write.format("parquet").saveAsTable(permanent_table_name)

df1 = table("cadastro_nomes_sexo")
df2 = spark.sql("select * from cadastro_nomes_sexo")

type(df1)

type(df2)

df1.head(5)

df2.take(5)

df1.show(5)

df2.count()
