==> 72-modelo-genero-nome-autokeras.ipynb <==
import pandas as pd
import numpy as np
import math

### Carga dos dados

# definir arquivo Pickle de entrada
#arquivo = 'dados/afastamentos-nomes-sexo-preparado.pkl.bz2'
arquivo = 'dados/cadastro-nomes-sexo-preparado.pkl.bz2'
#arquivo = 'extra/nome_sexo_pf-preparado.pkl.bz2'

df = pd.read_pickle(arquivo)
df.head()

df.shape

df.dtypes

### Expansão dos dados agregados

%%time
df2 = df.head(50000).copy()
#df2 = df.copy()
menor = df2['QTDE'].min()
maior = df2['QTDE'].max()
print(df2.shape[0], menor, maior, math.ceil(np.sqrt(maior / menor)))
#df2['TAM'] = df2['QTDE'].apply(lambda x: math.ceil(np.sqrt(x / menor)))
df2['TEMP'] = df2[['SEXO', 'QTDE']].apply(
    lambda x: [x[0]] * math.ceil(np.sqrt(x[1] / menor)), axis=1)
dfe = df2.explode('TEMP').drop(['TEMP'], axis=1) #.reset_index()
dfe.head(1000)

# usar os dados expandidos invés dos agregados
dfo = df # guardar o dataframe original
df = dfe # apontar para o dataframe expandido

### Uso dos dados originais

X = df.drop(['SEXO', 'QTDE'], axis=1) # tudo, exceto a coluna alvo
y = df['SEXO'] # apenas a coluna alvo
print('Forma dos dados:', X.shape, y.shape)

# dividir dados em treino, validação e teste

from sklearn.model_selection import train_test_split

TRAIN_RATIO = 0.75
VALIDATION_RATIO = 0.15
TEST_RATIO = 0.10

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=(1 - TRAIN_RATIO))

X_val, X_test, y_val, y_test = train_test_split(
    X_test, y_test, test_size=(TEST_RATIO / (TEST_RATIO + VALIDATION_RATIO)))

print("Shapes:\n"\
      "   Train      - X: %s, y: %s\n" \
      "   Validation - X: %s, y: %s\n" \
      "   Test       - X: %s, y: %s" %
      (X_train.shape, y_train.shape, X_val.shape, y_val.shape, X_test.shape, y_test.shape))

# dividir dados em treino e teste

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.30, random_state=42)

print("Shapes:\n"\
      "   Train      - X: %s, y: %s\n" \
      "   Test       - X: %s, y: %s" %
      (X_train.shape, y_train.shape, X_test.shape, y_test.shape))

X_train.head()

y_train.head()

import os
os.environ['PYTHONHASHSEED']=str(1)

import tensorflow as tf
import random
import numpy as np

def reset_random_seeds(SEED = 42):
  os.environ['PYTHONHASHSEED']=str(SEED)
  tf.random.set_seed(SEED)
  np.random.seed(SEED)
  random.seed(SEED)

import tensorflow as tf
import autokeras as ak

# Initialize the structured data classifier
clf = ak.StructuredDataClassifier(
    overwrite=True,
    max_trials=10 # It tries 3 different models
)
clf

X_train.to_numpy().astype('int8')

y_train.to_numpy().astype('int8')

%%time

# Feed the structured data classifier with training data
clf.fit(
    X_train.to_numpy().astype('int8'),
    y_train.to_numpy().astype('int8'),
    validation_split=0.15, # Split the training data and use the last 15% as validation data
    #epochs=10, # leave the epochs unspecified for an adaptive number of epochs
)

# evaluate the model
loss, acc = clf.evaluate(X_test, y_test, verbose=0)
print('Acurácia: %.2f %%' % (acc * 100))

model = clf.export_model()
model.summary()

type(clf)

type(model)

model = clf.export_model()

#print(type(model))  # <class 'tensorflow.python.keras.engine.training.Model'>

'''try:
    model.save("model_autokeras", save_format="tf")
except Exception:
    model.save("model_autokeras.h5")'''

print(type(model))

# save model to file
#model.save('modelos/modelo-genero-nome-autokeras.h5py')
model.save('modelos/modelo-genero-nome-autokeras', save_format='tf')
#model.save('modelos/modelo-genero-nome-autokeras.h5')

# carregar um modelo previamente gravado em disco
from tensorflow.keras.models import load_model

model = load_model('modelos/modelo-genero-nome-autokeras', custom_objects=ak.CUSTOM_OBJECTS)
#model = load_model("model_autokeras", custom_objects=ak.CUSTOM_OBJECTS)
#model = load_model('modelos/modelo-genero-nome-autokeras.h5')
model

df2 = X_test.copy()
df2['PROB'] = model.predict(X_test)
df2['REAL'] = y_test
df2['PREV'] = df2['PROB'].apply(lambda x: 1 if x > 0.5 else 0)
df2['DIF'] = df2[['REAL', 'PREV']].apply(lambda x: abs(x[0] - x[1]), axis=1)
#df2['REAL'].apply(lambda x: 'F' if x else 'M')
#df2['PREV'].apply(lambda x: 'F' if x else 'M')
df2.head(15)

df2[df2['DIF'] > 0].head(15)
